[refactor/#154] EventCategoryBar 공통 컴포넌트로 분리
