+++ app/src/main/java/com/bongtu/baekseo/presentation/record/RecordScreen.kt
@@ -24,13 +24,13 @@ import com.bongtu.baekseo.core.common.type.AttendType
 import com.bongtu.baekseo.core.common.type.EventCategoryType
 import com.bongtu.baekseo.core.common.type.ScheduleType
 import com.bongtu.baekseo.core.designsystem.component.BongBaekScheduleEmptyContent
+import com.bongtu.baekseo.core.designsystem.component.topbar.BongBaekCategoryBar
 import com.bongtu.baekseo.core.designsystem.theme.BongBaekTheme
 import com.bongtu.baekseo.data.model.event.ScheduleEvent
 import com.bongtu.baekseo.presentation.record.RecordContract.RecordSideEffect.NavigateToAdd
 import com.bongtu.baekseo.presentation.record.RecordContract.RecordSideEffect.NavigateToDetail
 import com.bongtu.baekseo.presentation.record.RecordContract.RecordUiState
 import com.bongtu.baekseo.presentation.record.component.AttendTypeTab
-import com.bongtu.baekseo.presentation.record.component.EventCategoryBar
 import com.bongtu.baekseo.presentation.record.component.RecordListContent
 import com.bongtu.baekseo.presentation.record.component.RecordTopBar
 import kotlinx.collections.immutable.persistentListOf
@@ -129,10 +129,11 @@ private fun RecordScreen(
             isEnabled = !uiState.isDeleteMode,
         )
 
-        EventCategoryBar(
+        BongBaekCategoryBar(
             selectedCategory = uiState.eventCategoryType,
             onCategoryClick = onCategoryClick,
-            isEnabled = !uiState.isDeleteMode,
+            isEnabled = true,
+            modifier = Modifier.padding(vertical = 20.dp),
         )
 
         Crossfade(
